backend/app.py
from flask import Flask, request, jsonify
from flask_cors import CORS
import os
from dotenv import load_dotenv
import cv2
import numpy as np
from PIL import Image
import io

# Load environment variables
load_dotenv()

app = Flask(__name__)
CORS(app)

# Initialize MongoDB connection (to be implemented)
# mongo_client = MongoClient(os.getenv('MONGODB_URI'))

@app.route('/api/health', methods=['GET'])
def health_check():
    return jsonify({"status": "healthy", "message": "FashionAI API is running"})

@app.route('/api/virtual-try-on', methods=['POST'])
def virtual_try_on():
    try:
        # Get user image and clothing item from request
        user_image = request.files.get('user_image')
        clothing_item = request.files.get('clothing_item')
        
        if not user_image or not clothing_item:
            return jsonify({"error": "Missing required images"}), 400

        # Convert images to numpy arrays
        user_img = Image.open(user_image)
        clothing_img = Image.open(clothing_item)
        
        # TODO: Implement virtual try-on logic using computer vision
        # This is where we'll integrate the AI model for virtual try-on
        
        return jsonify({
            "message": "Virtual try-on processing initiated",
            "status": "success"
        })

    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route('/api/style-recommendations', methods=['POST'])
def get_style_recommendations():
    try:
        data = request.json
        user_preferences = data.get('preferences', {})
        
        # TODO: Implement style recommendation logic
        # This is where we'll integrate the ML model for recommendations
        
        return jsonify({
            "recommendations": [
                {
                    "outfit_id": "1",
                    "confidence_score": 0.95,
                    "items": ["item1", "item2", "item3"]
                }
            ]
        })

    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route('/api/outfit-matching', methods=['POST'])
def outfit_matching():
    try:
        data = request.json
        items = data.get('items', [])
        
        # TODO: Implement outfit matching logic
        # This is where we'll calculate compatibility scores
        
        return jsonify({
            "compatibility_score": 0.85,
            "suggestions": ["suggestion1", "suggestion2"]
        })

    except Exception as e:
        return jsonify({"error": str(e)}), 500

if __name__ == '__main__':
    app.run(debug=True, port=5000) 