requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.2
opencv-python==4.5.3.56
tensorflow==2.7.0
pillow==8.3.2
pymongo==3.12.0
python-dotenv==0.19.0
scikit-learn==0.24.2
requests==2.26.0
gunicorn==20.1.0
